Add source location support to handler tracking.
The ASIO_HANDLER_LOCATION((file_name, line, function_name)) macro may
be used to inform the handler tracking mechanism of a source location.
This macro declares an object that is placed on the stack.

When an asynchronous operation is launched with location information, it
outputs lines using the <action> 'n^m', prior to the 'n*m' line that
signifies the beginning of the asynchronous operation. For example:

    @asio|1589341941.203368|6^7|in 'async_write' (./../../../include/asio/impl/write.hpp:330)
    @asio|1589341941.203368|6^7|called from 'asio::awaitable<void> echo(asio::ip::tcp::socket)' (coroutines_ts/echo_server.cpp:39)
    @asio|1589341941.203368|6*7|socket@0x55817c74aec8.async_send

If std::source_location or std::experimental::source_location are
available, the use_awaitable_t token (when default-constructed or used
as a default completion token) will also cause handler tracking to
output a source location for each newly created asynchronous operation.
A use_waitable_t object may also be explicitly constructed with location
information.